+++ jetbrains-core/resources/META-INF/ext-python.xml
@@ -28,7 +28,7 @@
     </extensions>
     <extensions defaultExtensionNs="aws.toolkit.codewhisperer">
         <programmingLanguage implementation="software.aws.toolkits.jetbrains.services.codewhisperer.language.languages.CodeWhispererPython"/>
-        <fileReader implementation="software.aws.toolkits.jetbrains.services.codewhisperer.language.classreader.CodeWhispererPythonReader"/>
+        <classResolver implementation="software.aws.toolkits.jetbrains.services.codewhisperer.language.classresolver.CodeWhispererPythonClassResolver"/>
         <importAdder implementation="software.aws.toolkits.jetbrains.services.codewhisperer.importadder.CodeWhispererPythonImportAdder"/>
     </extensions>
 </idea-plugin>
